Add the files and scripts for creating portable containers
This PR adds portable containers to enable developers who might not want to clone the entire repo + submodules and get all the dependencies. With these containers, it will now be possible to use a simplified docker-compose file or simply call docker `run` anywhere.

Our containers for the api are already very portable, but they needed to be sent to dockerhub during merges and PRs and to have the config files added at build them.