Make sure to filter the commit with the default branch
The previous code had a small issue where we would take the coverage from the latest commit, regardless of which branch that commit was from. This was an issue since a repository should only display the coverage of the main default branch.

This PR adds a small parameter to the subquery and adds a test to make sure it works as expected.