Add a new query parameter to order charts by commit dates
Depending on the use case, it would be useful to be able to get the coverage entries of the chart endpoint in increasing or decreasing order. Due to a misunderstanding of the code and the results, the order was changed to decreasing, which was not the intended behaviour in the base use case.

To make sure we support all use cases, I added a configurable parameter that can be added to the body of the request to decide the order of the coverage data based on the commit dates.

I also added unit tests so we can be informed of any unintended change of this behaviour in the future.