Add a truncated version of the timestamp for filtering
When we were filtering the commits by date, we used the database timestamp. In the case of the endpoints with by the analytics page, that timestamp filtering does not have the time data. This led to a situation where the code would think that a commit with a timestamp at 1AM was greater than the timestamp given at 0AM and we would get the previous commit instead of the one we wanted.

This PR adds an annotated truncated date function that will allow us to filter by a date with the time data omitted.

Work is related to https://codecovio.atlassian.net/browse/CE-2284