Porting Github oauth
Porting Github Login view from legacy.

It is a standard oauth app. When the user first hits the url, we redirect them to Github, while telling (via url args) who we are and what permission scope we want.

Then Github redirects back, with a 'code' argument on the url. We can use that code to fetch an access token for the user. From there, we can use the normal torngit interface.

This view is being ported without the backfills yet. After we do all providers, we can decide what to put there