Experimenting with pip-tools
Doing what we did at the worker: https://github.com/codecov/worker/pull/330

This works pretty straighforwardly. No one should touch the requirements.txt file directly. We all update the requirements.in file when we need to change/add a dependency and run:

```
pip-compile requirements.in
```

pip-tools already takes care of finding the requirements that will match all first-level requirements we have. And the end-result can be simply used normally on deployments and docker as we already use.

I left the original value exactly as it is (except for replacing a branch with a tag from shared that I mistakenly left there), for the least possible impact. But ideally we should have less strict versions on requirements.in (read major.minor but not major.minor.micro). But that's up to taste