Catching Torngit errors on login
Catching Torngit errors on login on Github and Gitlab.

On GH and GL, the only time we explicitly talk to the provider is after they redirected the user back to us (oauth2).

Right now we don't have a good place to redirect to. So we are setting it as redirecting back to the home page and logging the issue. I believe this is better than the current behavior of just showing an error message.

Hopefully we can soon get a special page for it. Then we can just replace the `error_redirection_page` with the name of that page