Rerolling Django secret key
According to https://docs.djangoproject.com/en/3.2/ref/settings/#std:setting-SECRET_KEY the secret key is used for:

```
The secret key is used for:

    All sessions if you are using any other session backend than django.contrib.sessions.backends.cache, or are using the default get_session_auth_hash().
    All messages if you are using CookieStorage or FallbackStorage.
    All PasswordResetView tokens.
    Any usage of cryptographic signing, unless a different key is provided.

If you rotate your secret key, all of the above will be invalidated. Secret keys are not used for passwords of users and key rotation will not affect them.
```

None of those things are things we use actively in our system.

Generated new key using

https://github.com/django/django/blob/0c71e0f9cfa714a22297ad31dd5613ee548db379/django/core/management/utils.py#L77-L82

which is the default way to generating the secret key in the first place.

We could theoretically put this value in the secrets manager. But I think until we start using the above features, it might just make it more confusing on some stuff