Adding database columns to pulls table
Adding column `user_provided_base_sha` to the `pulls` table.

The idea is that this is a field that the customer can set to tell what `base` they want to be used.

Unfortunately, we can't just use the `base` column, because it's overwritten by our own calculations all the time (we could eventually rename it to `calculated_base`, though.