Profiling app and models
Writing what the profiling models would look like. The two entities are:

`ProfilingUpload` - Each upload made by the user. This is the object that we expect to appear a lot of times, since some customers might make a lot of uploads. It has a `raw_upload_location`, which is basically the storage filepath for where we tell the client to upload the profiling data to (so we don't have to deal with big uploads ourselves)

`ProfilingCommit` - This should contain all the profiling information pertaining to a particular commit (even though on this new version a commit SHA is not really necessary and won't be used, and the user can just use the `version_identifier` field). This will be the grouper to all the uploads relating to the same code version.

Every once in a while (in a crontab) we will go through a lot of `ProfilingUpload`, group them by `ProfilingUpload`, then gather their data and join them into a single "per commit profiling data" file and save it on a location on storage. Then we save that location to `ProfilingCommit.joined_location`. On a second step, we will be summarizing that data and storing the summarization result on `ProfilinfCommit.summarized_location`.

The reason on why we have both locations separate is that both totals (which contain all details), and summary (which does not contain all details but is smaller and easier to use on different places) is that both have their place on the analysis. Also, there might be a point in the future where we don't join data but instead gather it all from a secondary place. In that case the joining logic changes, but the summarization stays the same.

Among its other main fields is `ProfilingCommit.last_joined_uploads_at`: it's meant for registering up to which point were the `ProfilingUpload`s were joined joined. Since this is not sensitive to all the user-made uploads (meaning if we miss a single upload that's fine), we can just choose the time windows of ProfilingUploads we already joined into a "per commit profiling data" archive.

Alternatively we can just set a boolean "was_already_joined" on the `ProfilingUpload`. We are analyzing this possibility, but the fear is that it is a lot of individual pieces to keep managing and updating. There is no intrinsic reason why it wouldn't work, but for now the time-window based analysis seems enough.