Picking proper service_id for Bitbucket
We were previously picking

```
user_data.pop("account_id")
```

as the service_id for a new user on Bitbucket, instead of the proper

```
user_data["uuid"][1:-1]
```

This lead a few things:

- Since the service_id, even for a returning user, was a new one, a new `Owner` was created. This is
- The new `Owner` had the same service/username as the old `Owner`, but the wrong service_id.
- This lead to a very particular database trigger to kick in (this is from before we got here):

```
create or replace function owners_before_insert_or_update() returns trigger as $$
  begin
    -- user has changed name or deleted and invalidate sessions
    with _owners as (update owners
                     set username = null
                     where service = new.service
                       and username = new.username::citext
                     returning ownerid)
      delete from sessions where ownerid in (select ownerid from _owners);
    return new;
  end;
$$ language plpgsql;
```

which would rename the username on the right-serviceid-owner to `null`.

= Since this was a login, the `SyncTeamsTask` would kick in. On that SyncTeamsTask, one of the synced teams would be the old right-serviceid-owner (because it was found via its `service_id`), the right-serviceid-owner would then get back the right username, and the database trigger would again kick in to set the username on the wrong-serviceid-owner to NULL.

So this whole username-renaming and session deleting would break the flow from the users, specially given it depends on the username for logins.

So this PR fixes that. It also adds some tests to ensure the right uuid is used and it re-enables and old test that needed to be skipped in the past