Removing `commit_sha` column for now
We will need this commit_sha eventually so we can get adjust the profilings with the diff to get the critical lines with correctness.

This is also relevant with file-level criticality (for file renaming). But less relevant, so this is going to happen only after the rest of the features is implemented. So we don't need it right now

Until we decide we need to use the `commit_sha` for something, let's leave it out of the database.

So when we do need it, we do it the right way.