Removing create_root_storage calls
This is a call that is not meant to be called on every moment. Only at code startup.

Its purpose is to create buckets for the system that is just installing codecov. It is otherwise a no-op.

On production, we don't need it. On enterprise, worker is already doing that for new customers

So removing that call and its mocks