Creating profiling views
Creating views for allowing the users to send the profiling data

There are also a couple of other assumptions/decisions here:
1. If at the point of upload, the customer chooses an version identifier that does not have a corresponding `ProfilingCommit` at the database yet, we just create it for them, instead of raising a 404
    - The rationale for that is personal experience with sentry sdk which has similar behavior. The effort to call that endpoint explicitly on the end of CI is a relatively big lift for us, so we don't so it. Their UI complains to you that the current version was not explicitly created via its own special API call, but it creates them and still allows you to do most things you need there.

2. The url of the view is currently

```
/profilings/uploads
```

I think there are alternatives here, ranging from having the repository at the url, to having nothing to having just the profiling_commit identifier in the url. Having the commit is a little iffy because the "calls might be made from the customer's customer's browser might require them to somehow give users information about their repository names.

Having the profiling_commit identifier in the url might work, if that's better than the current view

Regardless, the system finds the proper repository from the authentication token.

3. There are two new columns at the `ProfilingUpload` model: normalized_at and normalized_url. Those are fields that are used to do the conversion between the user-inputted value, and the value we store internally. This is meant to ease the collection of those uploads later.

4. We create presigned puts for the customer to upload all their data to the storage system directly. This eases the load on our system.