Creating Repository Token Authentication
This change creates the `RepositoryToken` model.

This model is heavily inspired on the DRF token authentication system.

The user is able of creating different tokens for particular repository scopes (such as report uploads, or `profiling` uploads). The idea is that by creating different scopes for different capabilities, the user has limited impact on possible leaks. And this is specially useful on situations where the user is running codecov from places other than their own machines (such as running profiling on their user's browsers).

The authentication itself is the usual token one: a header where the user has a prefix and the token.