Setting `using_integration=False` on trigger_refresh
`using_integration` is not really intuitive. What it really does is, as described better on legacy ( https://github.com/codecov/codecov.io/blob/v4.4.9/app/tasks/refresh.py#L45 ) is to sync which repos of a given owner use the integration.

It's not meant for a "this owner has integration installed" switch on other syncs. In fact, in legacy, the only times this gets called with `using_integration=True` is on `('installation', 'integration_installation', 'integration_installation_repositories')` webhooks.

So this needs to be set to False, because if it is True, it blocks the users ability to sync properly on the UI.

It is not a blocker for most people, since the login sync still seems to properly use the `using_integration=False` parameter