Removing asyncio runs
It looks like a given page is simply hanging when getting to a piece of logic near an asyncio.run

This corroborates with the fact that `asyncio.run` at Django code might make the request hang. We have previously seen similar code dying near an asyncio.run at the upload endpoint.

We still don't know if that is the cause. But regardles, we are replacing all asyncio.run from `services.comparison` to see if that indeed fixes something.

If it does, it might be worth it to go around the code and replace all asyncio.run in the api.