Changing iexact to exact on django admin
On Django admin for the repos page we have a way to search by owner.username.

The field for that is:

```
author__username__iexact
```

But when using that, the query produce has a `WHERE UPPER("owners"."username"::citext) = UPPER('codecov')`. On explain, we see that this is not able to use the owners index:

```
thiago-ro@34:postgres> EXPLAIN SELECT * FROM owners WHERE UPPER("owners"."username"::citext) = UPPER('codecov')
+---------------------------------------------------------------+
| QUERY PLAN                                                    |
|---------------------------------------------------------------|
| Seq Scan on owners  (cost=0.00..47431.26 rows=6278 width=889) |
|   Filter: (upper((username)::text) = 'CODECOV'::text)         |
+---------------------------------------------------------------+
```

Changing it to

```
author__username__exact
```

produces a slightly different query, which is able to use the index:

```
thiago-ro@34:postgres> EXPLAIN SELECT * FROM owners WHERE username='codecov';
+-------------------------------------------------------------------------------------------+
| QUERY PLAN                                                                                |
|-------------------------------------------------------------------------------------------|
| Index Scan using owner_service_username on owners  (cost=0.43..30257.57 rows=1 width=889) |
|   Index Cond: (username = 'codecov'::citext)                                              |
+-------------------------------------------------------------------------------------------+
```

I believe this will make the queries faster