Creating users-basic plan
(to be merged on the maintainance window)

This will run

```
ALTER TYPE plans ADD VALUE 'users-basic';
```

Which will add a new possible type to the plans enum

Internet says it is zero-downtime. But we are setting a maintenance window to be sure.

Had to add `atomic=False` because the migration will not run inside a transaction

This feature seems to have been added on postgres 9.1 . Ours is 9.6. 91 is not even a supported postgres version anymore, so if enterprise customers have something before it, it's a problem but we should consider recommending upgrading for other reasons