Replacing custom rolled exception catching with get_or_create
Currently, on the `insert_commit` logic, we try to get a commit, then create it if it doesn't work

Which is usually a normal way of doing this kind of thing.

The problem is that sometimes there is a race condition where two commits try to call the create at the same time, which leads to an IntegrityError. And it seems that either postgres or SQL sidecar is not happy with it, and keeps dropping the connection when that happens.

We are unsure if anything can be configured to tell postgres to just not worry too much when that happens

But what we want to try (and is very likely that won't change much) is to call Django's `get_or_craete` instead.

https://github.com/django/django/blob/3.1.13/django/db/models/query.py#L563

It has some logic that wraps up the create part in an `transaction.atomic` block, which properly commits and rollbacks as needed.

Running the create part inside a transaction could be what the database needs to avoid this crashing. But it is unlikely because I think we did this in the past on a different query and it didn't work