Reindenting and reorganizing the dockerfile
The endgoal is to remove this and start using the Dockerfile.requirements directly

But to do that, we need to see them being the same. So this change only re-indents file

The one actual change is on line 61 where we add a `--no-index` which Dockerfile.requirements (both on API and worker) already use.