Add YamlHistory model to codecov-api. (#954)
* Add YamlHistory model to codecov-api.

Apparently the table yaml_history is missing from fresh installs of enterprise.
Comes from here: https://github.com/codecov/codecov.io/blob/master/src/sql/main/tables/yaml.sql

That breaks some functionality so the migration needs to be added to codecov-api.
Eventually we plan to drop web completely from self-hosted instances, so a model has been created in codecov-api.
That way we can eventually implement the functionality that web has here.

* create yaml history model in legacy_migrations and separate state and db.

We make use of separate state and db to tell django to execute the sql code while recording a certain state.
Like so we can have a model (needed because of tests) and still run the sql we want.

The tests issue arises when django wants to flush the db. Because yaml_history references owner, django needs to know about both models.
So we are stuck having a model for yaml_history, even though it's legacy.

No change needed in the migrations apllication logic.