Creating Label Analysis Models
Label Analysis has a relatively simple model:

The customer will say "Given what you know about commit ABC (base), can you tell me which tests are possibly affected on commit CDE (head)?". Then we create a request instance with state == "created", and send to the worker to process. When the worker is finished, it sets the state to "complete", and saves the result on the `result` column

So the fields are:

base_commit - The commit that is the reference as the calculation
head_commit - The commit we want to know about
requested_labels - Which labels the customer is interested in knowing about. This will be useful if the customer has many separate testing suites, and for now they only need to know about one of the suites.
state_id - The state of the processing: it can be `created`, `finished` or `error` for now. This is read-only for the customer, will be set as started by default, and the worker will be the one changing it as it processes
result - The final result of the processing. This has a `json` type on purpose, so we can change what we return as we decide which things are relevant. For now we only are returning the list of labels on each categorization ("this label was affected by the diff", "this label is in the report but not affected by the diff", "this label is not in the report yet, so it might be new"

The generated migration SQL is

```
BEGIN;
--
-- Create model LabelAnalysisRequest
--
CREATE TABLE "labelanalysis_labelanalysisrequest" (
    "id" bigserial NOT NULL PRIMARY KEY,
    "external_id" uuid NOT NULL,
    "created_at" timestamp with time zone NOT NULL,
    "updated_at" timestamp with time zone NOT NULL,
    "requested_labels" text[] NULL,
    "state_id" integer NOT NULL,
    "result" jsonb NULL,
    "base_commit_id" bigint NOT NULL,
    "head_commit_id" bigint NOT NULL
);
ALTER TABLE "labelanalysis_labelanalysisrequest" ADD CONSTRAINT "labelanalysis_labela_base_commit_id_8cb5e96c_fk_commits_i" FOREIGN KEY ("base_commit_id") REFERENCES "commits" ("id") DEFERRABLE INITIALLY DEFERRED;
ALTER TABLE "labelanalysis_labelanalysisrequest" ADD CONSTRAINT "labelanalysis_labela_head_commit_id_58de868c_fk_commits_i" FOREIGN KEY ("head_commit_id") REFERENCES "commits" ("id") DEFERRABLE INITIALLY DEFERRED;
CREATE INDEX "labelanalysis_labelanalysisrequest_base_commit_id_8cb5e96c" ON "labelanalysis_labelanalysisrequest" ("base_commit_id");
CREATE INDEX "labelanalysis_labelanalysisrequest_head_commit_id_58de868c" ON "labelanalysis_labelanalysisrequest" ("head_commit_id");
COMMIT;

```