Implementing Label Analysis Views
Implementing the views for label analysis

The API itself is pretty simple, customers create the label request, and the system dispatches a task to the worker, which will analyze it.

Then a second API can be used to get details from the request, and whether it has completed

The request is:

```
{
    "base_commit": "1fbe7f5d509ac5721d970d9f4ff4019fcdb53797",
    "head_commit": "f21c4de0145183981d914421205abeefe9a3445c",
    "requested_labels": [
        "tests/test_eight.py::test_is_even_simple",
        "tests/test_eight.py::test_is_even",
        "tests/test_eight.py::test_is_even_really_good",
        "tests/test_eight.py::test_is_prime",
        // ...
        "tests/test_two.py::test_most_common_letters",
        "tests/test_two.py::test_letter_transformation"
    ]
}
```

The `requested_labels` is optional.

We might also consider in the future allowing a null `base_commit` and we will search in the database for suitable base commits (which have to be commits we have a static analysis and a report uploaded for)

And we respond with a very similar object, plus `state` and `result`