Don't pre-build session_totals (#1484)
Recent changes in `ReportFileSummary.session_totals` might be breaking because we pre-build them before callign the actual constructor ot `ReportFileSummary`. The new encoding format is a dict, not a list anymore. When iterating over a dict you get only the keys, but the values are lost.

We can leave for the constructor of `ReportFileSummary` to figure things out.