Extend model `ReportDetails` with multiple sources for `files_array`. (#1543)
* Extend model `ReportDetails` with multiple sources for `files_array`.

We want to move some of our data to GCS, the JSON fields. The first one, a bit of a POC
will be `ReportDetails.files_array`.

The present changes extend the model so that `files_array`, as a property, can come from
either the DB or storage. The property is responsible for deciding from where to get
the data.
Notice that this is a setup step, because we never save this data in api. A subsequent PR in
the worker will make us start using this.

We are using `@cached_property` so that we only have to get from storage once **per instance**.
Notice in test `reports/tests/test_omdels.py::ReportDetailsTests::test_get_files_array_from_storage`
that the last few asserts indicate the cache does not extends to different instances
of the same `ReportDetails` object.

Ticket: [CODE-3345](https://codecovio.atlassian.net/browse/CODE-3345)

* Combine migrations

Combining all DB operations in a single migration file

* Fix log extra. Decide on behavior for FileNotInStorageError