Extend API for label-analysis with update (#1585)
Update is PUT/PATCH for label analysis.
This change is an attempt to increase performance of ATS, and it's the first of 3 that need to happen.

ATS uses 3 important pieces of information to begin calculations:
1. labels collected;
2. the git diff;
3. the existing report for the base_commit.

All these informations are I/O bound from the point of view of ATS, and they all
can take considerable amount of time to be acquired (yes, [even the report](https://codecov.sentry.io/performance/python-3x-worker:8d9bd8626c0c46eea5cea7ca8312ebe0/?project=4165036&query=&referrer=performance-transaction-summary&statsPeriod=7d&transaction=app.tasks.label_analysis.process_request&unselectedSeries=p100%28%29)

With this change my hope is that we can have the CLI contact API twice, once before collecting the labels,
so that we can get started on fetching the diff and the report, and then another time to PATCH the request
with the labels being requested. Given that it can take many seconds to collect around 1000~2000 tests, this does
seem like a valid change.

The other 2 changes should happen in CLI and worker.