Create `GCSDecorator`
To make it easier for us to move more data to GCS in a transparent fashion we're introducing `GCSDecorator`.
It's not exactly a decorator, but you use `GCSDecorator.get_gcs_enabled_field` as the `fget` attribute of `property()` in the wrapped field you want and it works the same. `GCSDecorator.set_gcs_enabled_field` is for the `fset` attribute, obviously.

There's a bit of "inspection" (read hacky code) to access attributes in the object, as they can be different.
Particularly ugly is the `_follow_attributes_list`, that will allow us to get to the `repository` object (needed to create the `ArchiveService`) for the class. Notice that the path to get to it might differ from class to class, so we need to give it at declaration time (or implement the same path for every class, which is a ton of work).

One interesting side effect of doing this instead of using the previous definition is that the cache is propagated to
different instances of the same object (see `reports.tests.test_models.py::ReportDetailsTests::test_get_files_array_from_storage`).

Then why is it `default_value_fn` and not simply `default_value`? Well, to use with `property()` all attributes of the function need to be hashable. Lists are not hashable, but we use them quite a lot. However a Callable that takes no arguments is hashable, and we can use that to return whatever we want/need. That's the reason :)

For setting the attribute notice that we have `should_write_to_storage_fn`, so we can control what fields go to GCS and what fields don't.

You can also notice that there was some uplift in the Archive service so we can save data. All that code (including tests) was copied from the worker. I don't expect to use that a ton in the api, but if we need it, it's there. Being so transparent it's better to be safe than sorry.