Create `ArchiveField` (update of `GCSDecorator`)
Updates the `GcsDecorator` to `ArchiveField`. First change is the name. Because the archive service is not bound to GCS by any means ArchiveField is a better name. Also we're not using as a decorator, so it's better to not call it that.

THese changes are the result of suggestions made by @scott-codecov and @matt-codecov. Thanks for the suggestions :)

Now we are using a design pattern, the [Descriptor](https://docs.python.org/3/howto/descriptor.html). This makes it way better to understand what's going on and the implementation cleaner. It's also easier to use, since we don't have to rely on `property()` anymore. Just use it directly `files_array = ArchiveField(default_value=[])`.

It does come at it being way more opinionated about the class it's being used on. For example there's no more `attributes_to_repository`. Instead it's expected that the class that user `ArchiveField` implements `ArchiveFieldInterface`. There's actually a check on that when you actually use the `ArhiveField`, so that if the class doesn't implement the required interface it errors right out of the gate.

It's also opinionated on the field names. given a 'public_field' it's expected that the db field is '_public_field' and the archive one be '_publi_field_storage_path'. Interstingly checking `hasattr` on the field doesn't work. For some reason they don't appear in the attributes of the class... That made me scared that saving would fail, so I added the write test in `ReportDetails` to make sure. Apparently it works.