Remove race condition when saving data to storage
Moving the cache cleaning to after we update the field to prevent race conditions in which the cache is left in the old state.

Also adding deletion of orphaned data that was saved in GCS, in the event that the storage path changes.

Also moving the table name calculation to when we save the data in GCS, which is the only place we actually need it.