Move `Commit.report` to GCS
This is the API changes that will move `Commit.report` to GCS.
- Alter `report` field to be `_report` (the db_field) and `_report_storage_path` (the archive_field)
- Create the associated migration
- Turn `report` into an `ArchiveField`
- Adjust tests

There were some minor tweaks to `ArchiveField`, and I had to drop the typehint in `get_repository` to avoid circular dependencies
I also decided to move the config lookup of `should_write_data_to_storage` to a shared function, given they will be similar for all fields. Notice we can still turn individual fields on/off independently, but the "only_codecov" and "repo_ids" are shared now.

Every field still has their own `should_write_data_to_storage` to account for differences in the objects, and allow for individual behavior. Also less duplicated code, yay

Something else I had to do is include an `encoder` in the `ArchiveField` so we can continue using the same encoder for the db_field and the archive_field