Add trial logic (#30)
* Add trial logic

* Get rid of unecessary comments + linting

* reintroduce print statement and revert gql nullability

* Added expiration value when expiring trial

* Add pretrial user count field

* create a new endpoint that marks the completion step for uploads, which triggers notifs and syncs pulls (#49)

* fix: Cast UUID to str before saving in the db (#56)

* chore: add debug log for files array saving (#47)

In an effort to understand why we're getting `SessionArray` data from GCS without 'meta' info
we'll add a temporary log line when saving data from the API.
Mostly because I don't think this should happen, so I don't expect to see that line.
But who knows...

Part of codecov/platform-team#119

* chore: change logic to write data to storage (#48)

After moving `ReportDetails.files_array` and `Commit.report` to archive successfilly for Codecov
we want to do the same for some selected customers (in `repo_ids`) and then we will want
to start the process for new fields. Only for codecov.

The way things were before, if you have a field set to write to GCS it is valid for all codecov
repos AND all repos under `repo_ids`, while we would only want that for codecov repos.

So by changing the `master_write_switch` to strings we can have more power over each individual
field. Now we can use:
*  `True` or `codecov_access` to write *only* for codecov repos
*  `restricted_access` to write to all repos in `repo_ids` (+ codecov)
*  `general_access` to write to all repos

Because currently we're only writing for codecov repos the change will not affect any customers

Part of codecov/platform-team#100

* fix: Force owner login during pagination test (#58)

* merge

* Small fixes + addons

* Misc additions to the plan service + trial

* fix: Ensure we don't accidentally leak stack trace info (#59)

* Add pretrial_plan resolver

* Fix test

---------

Co-authored-by: Dana Yaish <98817625+dana-yaish@users.noreply.github.com>
Co-authored-by: scott-codecov <scott@codecov.io>
Co-authored-by: Giovanni M Guidini <99758426+giovanni-guidini@users.noreply.github.com>