fix: dont share default_values + log fix (#84)
Don't share default values. The present changes make it so we always
generate a new reference to a new object when returning the default value, rather than
the same pointer to all objects using `ArchiveField`.

In terms of logging reducing the level of the line that says both dbfield and
archive_field are None. It was spamming the logs