feat: trigger label analysis task after update
From https://github.com/codecov/worker/pull/21 we have a different way of handling updates for `LabelAnalysisRequest` objects.
This is important because we want to use these objects to extract metrics from ATS.

However to actually have the updated values we need to trigger the task again.
These changes trigger the task again :3

closes codecov/engineering-team#456