feat: tokenless auth (part I)
context: codecov/engineering-team#736

We want to accept tokenless uploads for forks of public repos.
This is a feature that is currently missing from the CLI and upload.

These changes are part I of that.
What it adds:
* `TokenlessAuthentication` - a new auth class that controls
  when a tokenless upload is accepted
* scaffolding `TokenlessAuth`
* `TOKENLESS_AUTH_BY_OWNER_SLUG` feature for internal testing
* The logic that will let us accept a tokenless upload

What is missing:
* Implementation of the methods in `TokenlessAuth`, so that it can
actually be used
* Hooking up the auth to existing endpoints.

Conditions to accept tokenless upload:
1. Request has `X-Tokenless` header, value is the fork repo slug AND
 `X-Tokenless-PR` header, value is the PR number
2. Request is to one of the new upload endpoints
3. The encoded repo (in the request path) exists AND is public repo
4. (temp) the repo is from an owner that can accept tokenless
5. The git service provider reports that the PR in `X-Tokenless-PR`
indeed exists and is from `X-Tokenless` repo to the encoded repo.