fix test that fails depending on run order
If you run
```
pytest \
 graphql_api/tests/test_impacted_file.py::TestImpactedFileFiltering::test_filtering_with_successful_flags \
 compare/commands/compare/interactors/tests/test_fetch_impacted_files.py::FetchImpactedFilesTest::test_impacted_files_filtered_by_flags_and_commit_comparison_for_pull
```

you'll see the 2nd test fails.

This is because of the PullID that apparently both tests create a pullID of 1.
It is strange because they are `TransactionTestCase` and that's supposed to reset the DB,
but God knowns what is going on.

Hard setting the pullID to different numbers solves the issue, but I'll be the 1st
to admit is not a pretty fix