feat: Add a coverage enabled column to repos (#340)
Create a column to determine if a user has coverage running. This differentiates from active because that is if any commit report has been uploaded, we need this to differentiate between coverage and bundle analysis.

This copies the current active values to the new column, can do this because there is no upload for bundle analysis yet.