fix: if upload moves commit to other branch, possibly update branch model head (#345)
that PR updates commit.branch but it doesn't touch the Branch model. Branch has a head and base field, and we may want to update them when moving a commit from one branch to another

as-is this PR will recompute the head of the destination branch, so if the commit being moved is newer than other commits on that branch, it'll become the new head. we may similarly want to fetch any branches that the commit being moved was the head of and recompute heads for those as well.