use ghapp when getting torngit adapter
Considers ghapp installations when deciding if a torngit adapter
is for a repo with integration or not.

If a ghapp installation exist then repo is considered to be using
integration if it's covered by the installation.
Otherwise previous behavior is used.

The test `test_get_adapter_sets_token_to_bot_when_user_not_authenticated`
includes `owner` as `None` when calling `RepoProviderService.get_adapter` so I'm assuming
that can happen, and changing the typehint in the function.

Other functions that were typehinted with `Owner` instead of `Optional[Owner]` are also
changed.