fix: fetch accurate branch head
This commit is a fix for the recent issue where
branch heads were all set to the same value each
time they were updated due to a bug in the Branch
django model.

This fix creates a helper function that checks for the 2
remaining innacurate commit shas that a large number of
branches were set to. If the current branch head is set to
one of those shas, we update the branch head by fetching
the latest commit on that branch from the db.

Note that we must use a raw sql query here to update the
branch object because using the django ORM to update it
would lead to the same issue we are trying to fix.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>