fix: logged in users can't see coverage bug
This bug caused a little incident in Codecov.

The actual issue is that we can't make DB queries syncronously from an
async context.

When getting the file content we have to get an adapter,
and that - since the GithubAppInstallation work - required a query to
the database to get the default GithubAppInstallation for an `owner`.
That was the actual error that caused the incident.

Error surfacing was terrible in that endpoint so I changed the `info` log
from an `exception` one so we can have a stacktrace.

To actually fix the error I actually created an async version of `get_adapter`.
I scanned the code, but it seems that this instance is the only place where we
actually use this.