Update Bundle analysis Schema and add GQL cleanup step (#408)
Make a teardown step for GQL request to clean up /tmp directory. This is needed for BA because we load the SQLite file (used as a storage for bundle stats) from GCS into the containers file system in the /tmp dir, we need to delete that file at the end when the entire GQL request is resolved.