fix: carefully assign ghapp names (#467)
* fix: carefully assign ghapp names

The configuration for ghapps need to be done carefully before they can
be used. And that process is _manual_.

However it's very simple for a user to simply go and install an app to
their organization and point it's webhooks to us. In this case we would
think it's the default app and break the app.

So now we update the app's names everytime we get a new webhook for it.
Only the configured default app may use the default name, all others,
unless they manually receive a custom name, will be 'unconfigured_app'.

Closes: https://github.com/codecov/engineering-team/issues/1445

* don't enforce single default app

Allows for multiple apps with the default name,
as long as they are configured.

Basically for an app to have a custom name (including the default one),
it needs to be configured. Otherwise it's moved to the "unconfigured_app" category.