make `is_configured` more rebust
Recently an enterprise customer had a ghapp instance of the default
app change names to `unconfigured_app` even though it is the default
app.

One possible explanation is that in the YAML the default app id is a string,
not a number. This small change is to catch that case.

The motivation is:
```
>>> 10 == 10
True
>>> 10 == '10'
False
>>> str(10) == str('10')
True
```