Warn when ghapps flip names
We had again a situation in staging that the default app flipped names to 'unconfigured_app'.
Not sure why exactly, but it's anoying and disruptive.

These changes emit a warning when the app flips name to unconfigured so we can track better
that situation (but we hope it doesn't happen much) and hopefully get better understanding
on why it's happening.

I also made a change to the `is_configured` function that an app unconfigured is never configured.