fix: use new tokenless protocol
if a request is creating a commit then we require
the name of the branch it's creating a commit on to
contain a colon, because that means that it won't
affect the coverage of the branches that already belong
to that repository. This will only work for public
repositories.

Subsequent API calls to the reports and uploads endpoints
will check the branch and visibility of the repo those
requests are targetting.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>