feat: trigger PR sync
It's possible with a combination of not having webhooks and temporary errors when
fetching info from the git provider that a PR will be left forever open.

One solution is to make the PR view trigger a sync_pull task in the background.
Notice that for the person that is actually viewing the PR it will probably be out of date,
but if they refresh the page it will eventually be updated.

closes: https://github.com/codecov/internal-issues/issues/513