Fix call to `commit_components` to use the right type
It was originally passing in a `User` object, but actually needed an `Owner` object.
Getting the commit was (silently) failing on `service not found` because a service
is not associated with a `User`.

I tested this locally and it seems to get the YAML components for the commit now.
Previously, it would default to the repo YAML, which would be unstable based on
whether the repo YAML is set (or updated during sync).