chore: reduce usage of PreProcess task
Originally I wanted to delete PreProcess task completely, but after reviewing
certain usage patterns I realized the task is necessary (for `empty-upload` and
`static-analysis` for example).

So the alternative is to reduce the amount of times it is called, to only be called
when we effectively create a new report. Similar to https://github.com/codecov/codecov-api/pull/804