feat: rework aggregate_test_results to use DailyTestRollup
this commit also adds the LastDuration field to the TestResult
GQL model. The reason we're using latest_run and not the updated_at
field of the object is because in the future we want to parse the
timestamp from the JUnit XML file so latest_run will be tied to that
and not the time at which we process the test results.