fix: add cursor logic to test analytics query
we change the query to only do ordering and get the entire set of tests
then we do the filtering in the application using a binary search,
because the result of the query is ordered by the ordering parameter
then the name and the binary search takes that into account.

The reason we have to do the filtering in the application is so we can
get the correct value for the totalCount.

Otherwise we would have one query complete the logic for getting
totalCount and another query actually doing the cursor filtering.