Store v2 reports in GCS instead of Redis
Worker has some dedicated code to move an uploaded report from Redis over to GCS.
Instead of having that code, we can just store the report into GCS in the first place.

It also forwards all the task arguments directly to the `Upload` task, in addition to storing it in Redis, to be forward compatible to some more planned upload processing improvements.