Allow erasing a foreign Repo
This changes the `EraseRepositoryInteractor` to allow erasing repositories belonging to other owners the `current_owner` has admin access to.

While doing so, this also extracts some duplicated code related to resolving `Owner`s and `Repository`s by name, and validating admin access.

The main part to allow this functionality (changing the GraphQL query) is left to a followup PR.