"Fix" errors calling `get_flag_names` on `None`
The recent switch to `get_flag_names` has caused a couple of "new" Sentry errors. Though they are not really "new", but rather just different from previous errors.

It seems like the existing code is passing `None` as Report in a bunch of places, and has seemingly always done so. The error now is just a different one.

Nonetheless, I paper over that error in some places by defaulting to `[]` instead on an empty Report. The code will still fail on a different line though.