Reorder trace annotations
I was very surprised that a ton of functions that have a trace annotation are not actually creating separate tracing spans.

It is highly probable that the trace annotation interacts with some other graphql or async related annotations in such a way that it does not work properly. Reordering it to be the last (aka inner-most) should hopefully solve this problem.