Make delete confirmation more lightweight
This nulls out the `get_deleted_objects`, which is used by the django admin to display a list of objects, and more importantly related objects being deleted.

As in particular with big owners/repos, actually building the list of all the related objects is unreasonably slow, so slow in fact that showing the confirmation dialog times out.

To not block deletes of deeply nested object hierarchies, this will just return empty lists instead.