fix(ta-gql): get testsuites
I made a change previously that made it so the testsuite column in
the dataframe returned by get_results became a list[str] instead of
str but I didn't update get testsuites to reflect that. This fixes
that.

also adds a test for the new V1 format that also verifies it populates
the testsuites correctly and fixes a bug with that format renaming a
column